api/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
autopep8==2.0.2
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
cffi==1.16.0
click==8.1.3
cryptography==41.0.5
distlib==0.3.6
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.104.1
filelock==3.12.0
Flask==2.2.5
Flask-Cors==3.0.10
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
motor==3.3.1
orjson==3.9.10
passlib==1.7.4
Pillow==9.5.0
platformdirs==3.5.1
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.1
pyasn1==0.5.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
pymongo==4.6.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pywatchman==1.4.1
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
typing_extensions==4.8.0
ujson==5.8.0
uvicorn==0.24.0.post1
uvloop==0.19.0
virtualenv==20.23.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==2.3.4
